Storing offset bug fixed
